+++ examples/hello_world/hello_world.py
@@ -13,37 +13,33 @@ class HelloWorld(app.Application):
         c = HelloWorldController(self).setup()
         self.add_controller("main", c)
 
-    def run(self):
-        self.get_controller("main").start()
-
-
-class HelloWorldController(controller.Controller):
-
-    def setup(self):
-        if self.view is None:
-            self._view = HelloWorldView(self.application, self)
-        self.view.setup()
-        return self
-
     def start(self):
-        self.view.show()
-
-    def stop(self):
-        self.view.close()
+        self.get_controller("main").start()
 
 
 class HelloWorldView(view.View):
 
     def setup(self):
+        # resize frame with window size
         self.grid(sticky="nsew")
-        self.grid_columnconfigure(0, weight=1)
-        self.grid_rowconfigure(0, weight=1)
+        self.configure_row(self, 0)
+        self.configure_column(self, 0)
+        # parent should also resize with window
+        self.configure_row(self.parent, 0)
+        self.configure_column(self.parent, 0)
         lbl = view.ttk.Label(self, text="Hello World!", font=self.FONT_LARGE)
         self.add_widget("label1", lbl)
-        lbl.grid(sticky="nsew")
+        lbl.grid(sticky="nsew", padx=10, pady=10)
 
+
+class HelloWorldController(controller.ViewController):
+
+    VIEW_CLASS = HelloWorldView
+
+    def update(self, observable):
+        pass
 
 if __name__ == "__main__":
 
     app = HelloWorld()
-    app.start()
+    app.run()