+++ julesTk/app.py
@@ -1,4 +1,6 @@
-"""Module managing starting and stopping the application"""
+"""Module managing starting and stopping the application
+
+"""
 
 from . import tk, ThreadSafeObject
 
@@ -18,7 +20,7 @@ class Application(ThreadSafeObject, tk.Tk):
         """All controllers attached to this application
 
         :return: A dictionary of the name - controller mapping
-        :rtype: dict[str, pyLabJackView.controller.Controller]
+        :rtype: dict[str, pyLabJackView.controller.BaseController]
         """
         return self._controllers
 
@@ -29,7 +31,7 @@ class Application(ThreadSafeObject, tk.Tk):
 
         :param name: Name of the controller
         :type name: str
-        :rtype: julesTk.controller.Controller
+        :rtype: julesTk.controller.BaseController
         """
         if not self.has_controller(name):
             raise KeyError("No controller registered under: {}".format(name))
@@ -52,7 +54,7 @@ class Application(ThreadSafeObject, tk.Tk):
         :param name: Name to register the controller under
         :type name: str
         :param controller: The controller to register
-        :type controller: julesTk.controller.Controller
+        :type controller: julesTk.controller.BaseController
         """
         if self.has_controller(name):
             raise KeyError("Already registered a controller under: {}".format(name))
@@ -73,24 +75,24 @@ class Application(ThreadSafeObject, tk.Tk):
         """Configures the application and loads at least one controller"""
         raise NotImplementedError
 
-    def start(self):
+    def run(self):
         """Start the application"""
         with self.lock:
             # prepare the application
             self.setup()
             # now start the application
-            self.run()
+            self.start()
         # in the main loop we wait, so we do not need a lock
         self.mainloop()
         with self.lock:
             # clean up
-            self.quit()
+            self.stop()
 
-    def run(self):
+    def start(self):
         """Everything to show and run the applciation"""
         raise NotImplementedError
 
-    def quit(self):
+    def stop(self):
         while len(self.controllers) > 0:
             name = self.controllers.keys()[0]
             self.get_controller(name).stop()